apps/worker/pyproject.toml
[project]
name = "banghack"
version = "0.1.0-dev"
description = "TikTok Live AI Co-Pilot Worker"
authors = [
    {name = "Dedy Prasetiyo", email = "dedy@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "TikTokLive>=6.4.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "edge-tts>=6.1.0",
    "websockets>=12.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "litellm>=1.50.0",
    "cartesia>=2.0.0",
    "edge-tts>=6.1.10",
    "httpx>=0.27.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "pyyaml>=6.0.1",
    "rapidfuzz>=3.9.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
